Get rid of devm_fwnode_get_[index_]gpiod_from_child()
# Lines starting with # will be removed from the cover letter. You can use
# them to add notes or reminders to yourself.

This drops the last uses of devm_fwnode_get_[index_]gpiod_from_child()
from the tree and drops the stubs implementing this API on top of
devm_fwnode_gpiod_get_index().

Note that the bulk of users were converted in 2019, the couple of LED
drivers are all that have remained.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>

# You can add other trailers to the cover letter. Any email addresses found in
# these trailers will be added to the addresses specified/generated during
# the b4 send stage.

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20220902-get_gpiod_from_child-remove-a62638849e91",
    "base-branch": null
  }
}