tpm: st33zp24: switch to using gpiod API
This patch switches the driver from legacy gpio API to the newer gpiod
API, so that eventually we can drop the former.

During conversion it was noted that the gpio handling code for SPI and
I2C variants duplicates each other, so it was moved into the core.

Also, it seems that the driver never assigned tpm_dev->io_lpcpd in the
past, so gpio-based power management was most likely not working ever.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>