nfc: s3fwrn5: fix order of freeing resources
Caution needs to be exercised when mixing together regular and managed
resources. In case of this driver devm_request_threaded_irq() should
not be used, because it will result in the interrupt being freed too
late, and there being a chance that it fires up at an inopportune
moment and reference already freed data structures.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>