nfc: st95hf: switch to using gpiod API
This switches the driver to use the newer gpiod API instead of legacy
gpio API that we want to retire.

Note that "enable" GPIO looks more like an active low reset line than
active high enable line (judging by the code in
st95hf_send_st95enable_negativepulse() where we briefly drive the line
low to reset the chip), but it is a topic for another set of patches if
we decide to rename it.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>