ARM: dts: omap3-n900: fix LCD reset line polarity
The LCD driver (panel-sony-acx565akm), when probing, starts with line
driven low, and then toggles it to high and keeps it there. Also, the
line is driven low when powering off the device, and ls released when
powering it back on. This means that the reset line should be described
as "active low" in DTS. This will be important when the driver is
converted to gpiod API which respects the polarity declared in DTS.

Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>